Moved HandJointName to ihmc-robotics-tools
