--- ihmc-robotics-toolkit/src/main/java/us/ihmc/robotics/partNames/FingerName.java
@@ -1,10 +1,0 @@
-package us.ihmc.robotics.partNames;
-
-public enum FingerName
-{
-   THUMB, INDEX, MIDDLE, PINKY;
-   
-   public static final FingerName[] threeFingerValues = new FingerName[] {THUMB, INDEX, MIDDLE};
-   
-   public static final FingerName[] fourFingerValues = new FingerName[] {THUMB, INDEX, MIDDLE, PINKY};
-}
