+++ robotiq-hand-drivers/src/main/java/us/ihmc/robotiq/model/RobotiqHandModel.java
@@ -2,7 +2,7 @@ package us.ihmc.robotiq.model;
 
 import us.ihmc.robotics.partNames.FingerName;
 import us.ihmc.avatar.handControl.packetsAndConsumers.HandModel;
-import us.ihmc.humanoidRobotics.communication.packets.dataobjects.HandJointName;
+import us.ihmc.robotics.partNames.HandJointName;
 import us.ihmc.robotics.robotSide.RobotSide;
 
 public class RobotiqHandModel implements HandModel
