+++ robotiq-hand-drivers/src/main/java/us/ihmc/robotiq/simulatedHand/SimulatedRobotiqHandsController.java
@@ -3,11 +3,11 @@ package us.ihmc.robotiq.simulatedHand;
 import controller_msgs.msg.dds.HandJointAnglePacket;
 import us.ihmc.communication.OldHandAPI;
 import us.ihmc.humanoidRobotics.communication.packets.dataobjects.HandConfiguration;
-import us.ihmc.humanoidRobotics.communication.packets.dataobjects.HandJointName;
 import us.ihmc.humanoidRobotics.communication.subscribers.HandDesiredConfigurationMessageSubscriber;
 import us.ihmc.log.LogTools;
 import us.ihmc.mecano.multiBodySystem.interfaces.OneDoFJointBasics;
 import us.ihmc.robotModels.FullRobotModel;
+import us.ihmc.robotics.partNames.HandJointName;
 import us.ihmc.robotics.robotSide.RobotSide;
 import us.ihmc.robotics.robotSide.SideDependentList;
 import us.ihmc.robotiq.model.RobotiqHandModel;
